+++ trilab_jpk_vat/models/jpk_vat7m_v2_report.py
@@ -381,6 +381,7 @@ class JpkReportV2(models.AbstractModel):
         v7m_report = self.env['jpk.vat.7m'].create(
             {
                 'version': '1-0E',
+                'technical_version': 'v2',
                 'year': report_date.year,
                 'month': report_date.month,
                 'cel_zlozenia': options.get('cel_zlozenia', 1),